move StageConfig back to config crate
