+++ crates/stages/src/config/mod.rs
@@ -1,5 +1,3 @@
 mod prune;
-mod stage;
 
 pub use prune::*;
-pub use stage::*;
